+++ src/utils/errors.ts
@@ -1,3 +1,6 @@
+import { ErrorObject } from 'ajv/dist/2020.js'
+import { Middleware } from 'koa'
+
 export class HttpError extends Error {
   public status: number
   public response: Response
@@ -13,18 +16,18 @@ export class HttpError extends Error {
   }
 }
 
-// export class ValidationError extends Error {
-//   public errors: ErrorObject | unknown[]
+export class ValidationError extends Error {
+  public errors: ErrorObject | unknown[]
 
-//   constructor(message: string, errors: ErrorObject | unknown[]) {
-//     super(message)
-//     this.name = 'ValidationError'
-//     this.errors = errors
+  constructor(message: string, errors: ErrorObject | unknown[]) {
+    super(message)
+    this.name = 'ValidationError'
+    this.errors = errors
 
-//     // Set the prototype explicitly to maintain correct instance type
-//     Object.setPrototypeOf(this, ValidationError.prototype)
-//   }
-// }
+    // Set the prototype explicitly to maintain correct instance type
+    Object.setPrototypeOf(this, ValidationError.prototype)
+  }
+}
 
 export class ConfigError extends Error {
   constructor(message: string) {
@@ -33,5 +36,20 @@ export class ConfigError extends Error {
 
     // Set the prototype explicitly to maintain correct instance type
     Object.setPrototypeOf(this, ConfigError.prototype)
+  }
+}
+
+export const handleErrors: Middleware = async (ctx, next) => {
+  try {
+    await next()
+  } catch (e) {
+    if (e instanceof ValidationError) {
+      ctx.response.status = 400
+      ctx.response.type = 'json'
+      ctx.response.body = {
+        message: e.message,
+        detail: e.errors,
+      }
+    } else throw e
   }
 }